+++ lib/src/components/NxTableRow/NxTableRow.tsx
@@ -7,7 +7,7 @@
 import React from 'react';
 import classnames from 'classnames';
 
-import {addPropsToChildren} from '../../internalUtil/childUtil';
+import {addPropsToChildren} from '../../util/childUtil';
 
 import {Props, propTypes} from './types';
 export {Props} from './types';
