+++ lib/src/components/NxTreeView/types.ts
@@ -7,7 +7,7 @@
 import {ReactNode} from 'react';
 import * as PropTypes from 'prop-types';
 
-import { TooltipConfigProps, tooltipPropTypesShape } from '../../internalUtil/tooltipUtils';
+import { TooltipConfigProps, tooltipPropTypesShape } from '../../util/tooltipUtils';
 
 export interface Props {
   onToggleCollapse?: (() => void) | null;
