+++ lib/src/components/NxTreeViewSelect/commonTypes.ts
@@ -7,7 +7,7 @@
 import {ReactNode} from 'react';
 import * as PropTypes from 'prop-types';
 
-import { requiredNullableString } from '../../internalUtil/customPropTypes';
+import { requiredNullableString } from '../../util/customPropTypes';
 
 export interface Option {
   id: string | null;
