+++ lib/src/util/childUtil.ts
