+++ lib/src/util/jsUtil.ts
