+++ lib/src/util/tooltipUtils.ts
